requirements.txt

python-telegram-bot==22.1
flask==3.1.1
aiosqlite==0.21.0
gunicorn==23.0.0
python-dotenv==1.0.0
